recommendations/requirements.txt
openai
pydantic
python-dotenv
requests
rapidfuzz